把放下面的牌的業務邏輯放到middleware裡
