+++ src/store/reducers/cards/index.ts
@@ -5,9 +5,7 @@ import {
 } from '../../../pages/mainpage/constants'
 import { Card } from '../../types'
 import { Action } from 'redux-actions'
-import _cloneDeep from 'lodash.clonedeep'
 import { DroppingDeckActionType } from '../../actionTypes/reducers'
-import { isCanPut_BelowDecks, } from '../../../game_logic'
 import { DnDTransData } from '../../../utils/DnDModule/models'
 
 const getDefaultTmpDecksState = () => {
@@ -23,7 +21,10 @@ const getDefaultTmpDecksState = () => {
   return tmpDeckCardData
 }
 
-const tmpDecksReducer = (state = getDefaultTmpDecksState(), action: Action<DnDTransData>) => {
+const tmpDecksReducer = (
+  state = getDefaultTmpDecksState(),
+  action: Action<DnDTransData>,
+) => {
   return state
 }
 
@@ -65,49 +66,20 @@ export const getDefaultDroppingDecksState = () => {
   return finishDeckCardData
 }
 
-
-
-const dragLastCardFromThisDraggable = (
-  cardMap: Map<string, Card[]>,
-  draggableId: string
-) => {
-  const cardsOfThisDropable = cardMap.get(draggableId)
-  const currLength = cardsOfThisDropable.length
-  return cardsOfThisDropable[currLength - 1].value
-}
-
-const dndDoneOnDroppingDeckArea = (
-  prevDeck: Map<string, Card[]>,
-  dndData: DnDTransData
-) => {
-  const { from, to } = dndData
-  const toCardsId = dragLastCardFromThisDraggable(prevDeck, to)
-  if (!isCanPut_BelowDecks(parseInt(from.dragItemId, 10), toCardsId)) {
-    return prevDeck
-  }
-
-  const newDecks = _cloneDeep(prevDeck)
-  const moveData = prevDeck.get(from.fromDroppableId)[from.dragItemIndex]
-  newDecks.get(to).push(moveData)
-  newDecks.get(from.fromDroppableId).splice(from.dragItemIndex, 1)
-
-  return newDecks
-}
-
 const droppingDecksReducer = (
   state = getDefaultDroppingDecksState(),
-  action: Action<Map<BelowCardDroppableArea, Card[]> | DnDTransData>
+  action: Action<Map<BelowCardDroppableArea, Card[]> | DnDTransData>,
 ) => {
-  const { type: doWhat, payload, } = action
+  const { type: doWhat, payload } = action
 
   switch (doWhat) {
     case DroppingDeckActionType.INIT_SWAPPED_CARDS:
       return payload as Map<BelowCardDroppableArea, Card[]>
     case DroppingDeckActionType.DND_DROPPING_DECK_CARD:
-      return dndDoneOnDroppingDeckArea(state, payload as DnDTransData)
+      return payload as Map<BelowCardDroppableArea, Card[]>
   }
 
   return state
 }
 
-export { droppingDecksReducer, tmpDecksReducer, finishDecksReducer }
+export { droppingDecksReducer, tmpDecksReducer, finishDecksReducer, }
